0508
1.导入数据调整，2，导出CSV功能 3.图片中心显示，

+++ DCTS/UI/NewDiningsForm.cs
@@ -46,7 +46,9 @@ namespace DCTS.UI
                 this.openAtDateTimePicker.Text = obj.open_at.ToString();
                 this.closeAtDateTimePicker.Text = obj.close_at.ToString();
 
-
+                long folername = obj.id / 1000;
+                string lcoalPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory + "\\data\\images\\location_" + ComboLocationEnum.Dining.ToString().ToLower() + "\\" + folername + "\\", obj.img);
+                pictureBox1.ImageLocation = lcoalPath;
             }
         }
 
@@ -148,7 +150,8 @@ namespace DCTS.UI
                             ctx.SaveChanges();
                             if (hasImg)
                             {
-                                string copyToPath = EntityPathConfig.LocationImagePath(obj);
+                                string copyToPath = EntityPathConfig.LocationDiningImagePath(obj);
+                                if (!File.Exists(copyToPath))
                                 File.Copy(imgFilePath, copyToPath);
                             }
                         }
@@ -272,6 +275,7 @@ namespace DCTS.UI
             if (openFileDialog1.ShowDialog() == System.Windows.Forms.DialogResult.OK)
             {
                 this.imgPathTextBox.Text = openFileDialog1.FileName;
+                   pictureBox1.ImageLocation = openFileDialog1.FileName;
             }
         }
